requirements.txt
# requirements.txt

# BDD Framework
pytest-bdd==7.0.1
behave==1.2.6

# Web Testing
playwright==1.40.0
selenium==4.16.0

# API Testing
requests==2.31.0
httpx==0.25.2
jsonschema==4.20.0
pydantic==2.5.0

# Data Testing
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
faker==20.1.0
great-expectations==0.18.5

# AI/ML Testing
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
deepeval==0.20.0
mlflow==2.9.0

# Performance Testing
locust==2.19.1
pytest-benchmark==4.0.0

# Reporting
allure-pytest==2.13.2
pytest-html==4.1.1
pytest-json-report==1.5.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2
tenacity==8.2.3

# Development
pytest==7.4.3
pytest-xdist==3.5.0  # Parallel execution
pytest-cov==4.1.0    # Coverage
black==23.12.0       # Code formatting
flake8==6.1.0        # Linting